+++ MissionSite/Models/Users.cs
@@ -11,7 +11,7 @@ namespace MissionSite.Models
     public class Users
     {
         [Key]
-        public int UserID { get; set; }
+        public decimal UserID { get; set; }
         public string UserEmail { get; set; }
         public string UserPassword { get; set; }
         public string UserFirstName { get; set; }
